+++ generate-requests.py
@@ -14,8 +14,9 @@ if __name__ == '__main__':
     parser = argparse.ArgumentParser(description='Connect to random hostnames with multiple threads.')
     parser.add_argument('--cleanup', action='store_true', help='Clean up downloaded files and exit.')
     parser.add_argument('num_connections', type=int, nargs='?', default=100, help='Number of connections to establish. Default 100.')
-    parser.add_argument('num_workers', type=int, nargs='?', default=10, help='Number of worker threads. Default 10.')
+    parser.add_argument('num_workers', type=int, nargs='?', default=3, help='Number of worker threads. Default 3.')
     parser.add_argument('--outputfile', type=str, help='Save output to a file.')
+    parser.add_argument('--insecure', action='store_true', help='Allow insecure connections.')
 
     args = parser.parse_args()
 
@@ -29,5 +30,5 @@ if __name__ == '__main__':
     file_manager.load_csv()
     file_manager.get_random_sample(args.num_connections)
 
-    connection_manger = ConnectionManager(num_connections=args.num_connections, num_workers=args.num_workers, outputfile=args.outputfile if args.outputfile else None)
+    connection_manger = ConnectionManager(num_connections=args.num_connections, num_workers=args.num_workers, outputfile=args.outputfile if args.outputfile else None, secure=not(args.insecure))
     connection_manger.main(file_manager.random_sample)